requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appscript==1.3.0
aspose-words==25.7.0
attrs==20.3.0
cachetools==5.5.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==41.0.7
distro==1.9.0
docx2pdf==0.1.8
Flask==2.2.3
fpdf==1.7.2
google-auth==2.40.3
google-genai==1.27.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
jiter==0.10.0
lib50==3.0.10
lxml==6.0.0
makelove==0.0.13
MarkupSafe==2.1.2
openai==1.97.1
packaging==23.2
panda==0.3.1
pexpect==4.9.0
pi==0.1.2
pillow==10.3.0
plum-dispatch==1.7.4
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.1
reportlab==4.4.3
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
submit50==3.1.3
tenacity==8.5.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
unicode==2.9
Unidecode==1.3.8
urllib3==2.1.0
websockets==15.0.1
Werkzeug==2.2.3
